make export_to_* methods be their own module, make export_to_csv doctest deterministic
